+++ extensions/config/portability-config-yaml/src/main/resources/META-INF/services/org.dataportabilityproject.config.extension.ApiSettingsExtension
@@ -1,1 +1,1 @@
-org.dataportabilityproject.config.yaml.YamlApiSettingsExtension
+org.dataportabilityproject.config.yaml.extension.YamlApiSettingsExtension
