+++ extensions/config/portability-config-yaml/src/main/resources/META-INF/services/org.dataportabilityproject.config.extension.WorkerSettingsExtension
@@ -1,1 +1,1 @@
-org.dataportabilityproject.config.yaml.YamlWorkerSettingsExtension
+org.dataportabilityproject.config.yaml.extension.YamlWorkerSettingsExtension
